pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = "setuptools.build_meta"

[project]
name='versionoverlord'
dynamic = ["version"]
description = 'Dependency Manager'
readme = "README.md"
license = {text = 'GNU AFFERO GENERAL PUBLIC LICENSE'}
authors = [{name = 'Humberto A. Sanchez II', email = 'Humbert.A.Sanchez.II@gmail.com'}]
maintainers = [{name = 'Humberto A. Sanchez II', email = 'Humbert.A.Sanchez.II@gmail.com'}]
keywords = ['version management', 'dependencies', 'python',]

dependencies = [
  'PyGithub>=2.5.0',
  'click~=8.1.7',
  'semantic-version==2.10.0',
]

[project.urls]
Repository = 'https://github.com/hasii2011/versionoverlord'


[tool.setuptools.packages.find]
where = ['src']

[tool.setuptools.package-data]
'versionoverlord.resources' = ['loggingConfiguration.json']

[tool.setuptools.dynamic]
version = {attr = 'versionoverlord.__version__'}

[project.scripts]

querySlugs          = "versionoverlord.commands.QuerySlugs:querySlugs"
createSpecification = "versionoverlord.commands.CreateSpecification:createSpecification"
updateDependencies  = "versionoverlord.commands.UpdateDependencies:updateDependencies"
versionOverlord     = "versionoverlord.commands.VersionOverlord:versionOverlord"
draftRelease        = 'versionoverlord.commands.DraftRelease:draftRelease'
qs = "versionoverlord.commands.QuerySlugs:querySlugs"
cs = "versionoverlord.commands.CreateSpecification:createSpecification"
ud = "versionoverlord.commands.UpdateDependencies:updateDependencies"
vo = "versionoverlord.commands.VersionOverlord:versionOverlord"
dr = 'versionoverlord.commands.DraftRelease:draftRelease'
